Stuff for the grid system.
Canvas.js contains a canvas object, instantiating a canvas creates a
canvas HTML element and inserts it into the index.html file that was also
included in this commit. The current width and height is based on the
Galaxy S4. Chose it over the iphone because I will not indoctrinated into
the Apple cult. Making it smaller to fit all phones might be better, that
is a simple change.

CanvasGenerator.js